1. change video 2. mute 3.full screen

+++ backend/src/main/java/com/quantil/webrtc/api/v1/meeting/dao/RtcMeetingItemDao.java
@@ -4,11 +4,11 @@ import com.quantil.webrtc.api.v1.meeting.bean.RtcMeetingItem;
 import java.util.List;
 
 public interface RtcMeetingItemDao {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(RtcMeetingItem record);
 
-    RtcMeetingItem selectByPrimaryKey(Integer id);
+    RtcMeetingItem selectByPrimaryKey(Long id);
 
     List<RtcMeetingItem> selectAll();
